- Refactoring to move out healthchecks and caching to seperate files. - Rename some methods for clarity. - Make code around eureka registration status clearer
